Moves "parseRequestBody" to the "utils/request" directory
